+++ weave/legacy/artifact_local.py
@@ -9,7 +9,7 @@ import tempfile
 import typing
 from datetime import datetime
 
-from weave import filesystem
+from weave.legacy import filesystem
 from weave import weave_types as types
 from weave.legacy import (
     artifact_fs,
