+++ weave/legacy/artifact_wandb.py
@@ -13,11 +13,9 @@ from wandb import Artifact
 from wandb.apis.public import api as wb_public
 from wandb.sdk.lib.hashutil import b64_to_hex_id, hex_to_b64_id
 
-from weave import (
-    filesystem,
-)
 from weave import weave_types as types
 from weave.legacy import (
+    filesystem,
     urls,
     errors,
     engine_trace,
