+++ weave/legacy/io_service.py
@@ -23,10 +23,10 @@ from typing import Any, Callable, Dict, Iterator, TypeVar
 import aioprocessing
 
 from weave import (
-    filesystem,
     weave_http,
 )
 from weave.legacy import (
+    filesystem,
     errors,
     engine_trace,
     server_error_handling,
